% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_deseq_res.R
\name{clean_deseq_res}
\alias{clean_deseq_res}
\title{Annotate arrange Deseq2 results}
\usage{
clean_deseq_res(deseq_res)
}
\arguments{
\item{deseq_res}{a Large Deseq Results}
}
\value{
clean_res_df: Deseq2 results which is DEG arranged ascendingly based on their P values
}
\description{
Function for adding q-values (FDR), annotate, arrange, and clean the deseq2 results.

Steps involved:
\enumerate{

\item Deseq2 results were converted into a data frame.
\item Row names were converted to the column (ENTREZID).
\item Q values (FDR) were calculated using the P-values and 'qvalue' package.
\item ENTREZID were Annotated to get the SYMBOL and GENENAME.
\item Annotations were merged with results.
\item The columns were rearranged and the results were arranged based on P values.

}
}
\examples{
clean_res <- clean_deseq_res(deseq_res)

}
