% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inflation.R
\name{inflation}
\alias{inflation}
\title{Calculate inflation factor}
\usage{
inflation(pvals)
}
\arguments{
\item{pvals}{Unadjusted P values}
}
\value{

}
\description{
The function calculates the inflation factor.
This function is part of the Q-Q plot generation and is needed by the \code{\link{rna_qq_plot}}.
}
\examples{
inflation(deg_df$pvalue)

}
