% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_prep.R
\name{pca_prep}
\alias{pca_prep}
\title{Prep for PCA plot}
\usage{
pca_prep(norm_dat, var_int, ntop = 1000)
}
\arguments{
\item{norm_dat}{data frame of normalized gene counts. Generally a made using the \code{\link{DESeq2}}.}

\item{ntop}{number of top genes that you need to calculate the PCs. Default is 1000 genes.}
}
\value{

}
\description{
pca_prep function uses the vst/log transformed data to make a dataframe that could be used to make a pca plot.
}
\examples{
pca_data_nb <- pca_prep(norm_dat = sample_vst_dd, "Group")

}
