% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna_pca_plot.R
\name{rna_pca_plot}
\alias{rna_pca_plot}
\title{PCA plot of the RNA samples}
\usage{
rna_pca_plot(pca_dat, x = 1, y = 2)
}
\arguments{
\item{pca_dat}{data frame that is the result after performing pca_prep function}

\item{x}{PC that should be displayed on the x axis. Default is 1, which is PC1.}

\item{y}{PC that should be displayed on the y axis. Default is 2, which is PC2.}
}
\value{

}
\description{
rna_pca_plot function makes a pca plot using the PC1 and PC2 to see how the samples are distributed.
}
\examples{
pca_data_nb <- pca_prep(norm_dat = sample_vst_dd, "Group")

# To plot PC1 and PC2
ana <- "Sample Title PC1 vs PC2"
caption <- "Data Source: \n Sample Data"

rna_pca_plot(pca_dat = pca_data_nb, x = 1, y = 2)

# To plot PC2 and PC3
rna_pca_plot(pca_dat = pca_data_nb, x = 2, y = 3)

}
