% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna_qq_plot.R
\name{rna_qq_plot}
\alias{rna_qq_plot}
\title{Quantile - Quantile plots for RNA seq}
\usage{
rna_qq_plot(pvals, ci = 0.95)
}
\arguments{
\item{pvals}{Un adjusted p values to make the QQ plot}

\item{ci}{Confidence interval. Default is 0.95.}
}
\value{

}
\description{
Function for creating the Quantile - Quantile plots of P values from DEG data frames
}
\examples{
caption <- "Data Source: \n Sample Data"
rna_qq_plot(sample_clean_deg$pvalue)


}
