% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna_vol_plot.R
\name{rna_vol_plot}
\alias{rna_vol_plot}
\title{Volcano plot for RNA DEG}
\usage{
rna_vol_plot(df, p = padj, val = 0.1)
}
\arguments{
\item{df}{data frame with differential gene expression with P values and Log fold changes}

\item{p}{pvalue (un adjusted) or padj (adjusted p value). Default is adjusted p value (padj).}

\item{val}{threshold to be colored. Default is 0.1.}
}
\value{

}
\description{
Function for creating the volcano plots from DEG RNA seq data frames.
Note: Default parameters would highlight the lo2foldchanges > |0.5|.
}
\examples{
# for using adjusted p values and highlight values less than 0.05
caption <- "Data Source: \n Sample Data"
rna_vol_plot(clean_deg, p = padj, val = 0.05)

# for using un - adjusted p values and highlight values less than 0.05
rna_vol_plot(sample_clean_deg, p = pvalue, val = 0.05)

}
