% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseqviz-package.R
\docType{package}
\name{rnaseqviz-package}
\alias{rnaseqviz}
\alias{rnaseqviz-package}
\title{rnaseqviz: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Kesava Asam \email{kra277@nyu.edu} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
